requirements.txt
streamlit
python-pptx
Pillow
